+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Updates
 
 ### Fixed
+- Inputmask numeric does no round up when digits is 0 #2018
 - Strange Calendar popup issue in IE Only when used with Daterangepicker #1965
 - incorrect work min max date - #2011, #2013